requirements.txt
pyvisa-py
pyserial 
pygpib
ThorlabsPM100